ci: lintcommit.js regressions #5515
Problem:
Since 8e34c5e650f3, `node lintcommit.js test` fails various tests and CI
fails for those cases too:

    /home/runner/work/aws-toolkit-vscode/aws-toolkit-vscode/.github/workflows/lintcommit.js:80
        } else if (!scope && typeScope.includes('(')) {

Solution:
Revert the changes from 8e34c5e650f3. It's good to avoid code
duplication, but in this case `parsePRTitle()` is not the right
abstraction:
- it doesn't signal failures in a way that is handled by callers
- its return type is awkward (`undefined | string | object`)
- notify.js doesn't actually need `parsePRTitle`, it only needs to check
`startsWith()`.
