+++ .github/workflows/lintcommit.js
@@ -11,7 +11,6 @@
 //
 
 const fs = require('fs')
-const { parsePRTitle } = require('./utils')
 // This script intentionally avoids github APIs so that:
 //   1. it is locally-debuggable
 //   2. the CI job is fast ("npm install" is slow)
@@ -69,7 +68,20 @@ void scopes
  * Returns undefined if `title` is valid, else an error message.
  */
 function validateTitle(title) {
-    const { type, scope, subject } = parsePRTitle(title)
+    const parts = title.split(':')
+    const subject = parts.slice(1).join(':').trim()
+
+    if (title.startsWith('Merge')) {
+        return undefined
+    }
+
+    if (parts.length < 2) {
+        return 'missing colon (:) char'
+    }
+
+    const typeScope = parts[0]
+
+    const [type, scope] = typeScope.split(/\(([^)]+)\)$/)
 
     if (/\s+/.test(type)) {
         return `type contains whitespace: "${type}"`
